changing folder structure and creating csv file
